temperatureカラムのデータ型をinteger → floatに変更

+++ db/migrate/20240314193014_create_temperatures.rb
@@ -2,7 +2,7 @@ class CreateTemperatures < ActiveRecord::Migration[7.0]
   def change
     create_table :temperatures do |t|
       t.time :time, null: false
-      t.integer :temperature, null: false
+      t.float :temperature, null: false
       t.string :memo
 
       t.timestamps
